requirements-freeze.txt
alabaster==0.7.16
-e git+https://github.com/ground-truth-ai/amlaidatatests.git@013225f8c7ad43e325c757d0e90a4d5910cf0f05#egg=amlaidatatests
-e git+https://github.com/ground-truth-ai/amlaidatatests.git@013225f8c7ad43e325c757d0e90a4d5910cf0f05#egg=amlsynth
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apeye==1.4.1
apeye-core==1.1.5
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
astroid==3.2.2
asttokens==2.4.1
atpublic @ file:///home/conda/feedstock_root/build_artifacts/atpublic_1641909357516/work
attrs==23.2.0
autodocsumm==0.2.12
Babel==2.15.0
beautifulsoup4==4.12.3
bidict @ file:///home/conda/feedstock_root/build_artifacts/bidict_1708298236075/work
black==24.4.2
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
CacheControl==0.14.0
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
coverage==7.5.3
cssutils==2.11.1
dacite==1.8.1
db-dtypes==1.2.0
decopatch==1.4.10
decorator==5.1.1
dict2css==0.3.0.post1
dill==0.3.8
docutils==0.21.2
domdf-python-tools==3.9.0
duckdb @ file:///home/conda/feedstock_root/build_artifacts/python-duckdb-split_1716489283545/work/tools/pythonpkg
execnet==2.1.1
executing==2.0.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1714422806336/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1707102468451/work
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
graphviz @ file:///home/conda/feedstock_root/build_artifacts/python-graphviz_1711016462626/work
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
humanize @ file:///home/conda/feedstock_root/build_artifacts/humanize_1700607927158/work
-e git+https://github.com/ibis-project/ibis@668af09904e740c6afc5e184bf4033e03acd491d#egg=ibis_framework
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
imagesize==1.4.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1711040877059/work
iniconfig==2.0.0
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1714665484399/work
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
makefun==1.15.2
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899920239/work
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
more-itertools==10.3.0
msgpack==1.0.8
multipledispatch==0.6.0
mypy-extensions==1.0.0
natsort==8.4.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225359967/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=031b7d6b2e5e604d9e21fc21be713ebf28ce133ec872dce6de006742d5e49bab
oauthlib==3.2.2
omegaconf==2.3.0
omegaconf-argparse==1.0.1
packaging==24.0
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1715897630316/work
parso==0.8.4
parsy @ file:///home/conda/feedstock_root/build_artifacts/parsy_1677097220753/work
pathspec==0.12.1
pexpect==4.9.0
pins @ file:///home/conda/feedstock_root/build_artifacts/pins_1715736000621/work
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1700596371886/work
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pylint==3.2.5
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==8.2.2
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373450623/work
referencing==0.35.1
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1715828414513/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-file==1.5.1
requests-oauthlib==2.0.0
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
rpds-py==0.19.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==70.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinx-autodoc-typehints==2.2.2
sphinx-data-viewer==0.1.4
sphinx-jinja2-compat==0.3.0
sphinx-needs==2.1.0
sphinx-prompt==1.8.0
sphinx-tabs==3.4.5
sphinx-test-reports==1.0.2
sphinx-toolbox==3.7.0
sphinxawesome-theme==5.2.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlglot @ file:///home/conda/feedstock_root/build_artifacts/sqlglot_1714501989735/work
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tomlkit==0.12.5
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1706112571092/work
traitlets==5.14.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717287769032/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1696486305723/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
